Add mutation testing with Stryker
Add StrykerJS(+typechecker) to run mutation tests on this project. This
type of testing introduces slight changes (mutants) into the source code
and runs the tests to check if this is detected. If not, the tests can
be improved.

A configuration file to enable running mutation tests was added. Running
mutation tests will generate files into two directories, both of which
are ignored by git. The mutation tests are also run in the CI.

The TypeScript configuration was updated to reflect files that should be
ignored by TypeScript when building. This is necessary for Stryker to
work.