Set up test coverage support
Add c8 to generate a coverage report for tests. Add an npm command to
run tests with coverage. Add a `.gitignore` entry to ignore the folder
that coverage reports go into. Update the CI to run tests with coverage
and publish the coverage report to Codecov, including a configuration
file for Codecov.