Create compatibility test suite
Create an initial compatibility test suite that tests whether the plugin
can be used with all version of ESLint it claims to support (in
`package.json` at (now) line 39).

The test suite invokes the ESLint CLI for each supported version (all
included as a devDependency with a custom name in `package.json`) with
the current plugin and rules configured on a variety of code snippets
and checks that the ESLint output (stdout + status code) are as expected
when compared to snapshots.

To reduce the scope of this commit, the compatibility test suite is
executed with the pre-existing tests for now.