Add 'better-npm-audit' to be able to ignore audit warnings
Add better-npm-audit as dependency to be used for auditing dependencies,
primarily to be able to ignore warnings if necessary. The `.nsprc` file
is the configuration file for better-npm-audit, as there are currently
no audit warnings it's still empty.

A shorthand for using better-npm-audit for auditing was added as an npm
command under the name "audit". This means that one should use `npm run
audit` over `npm audit` when auditing this project's dependencies.

The reusable-audit.yml workflow has been updated to run `npm run audit`
as well. For now, v0 will still run `npm audit` for compatibility
reasons (the v0 branch does not yet know about the "audit" command).