Add `exports` to the package manifest
Add the "exports" field [1] to the package manifest as the more modern
alternative to "main". Following [2], no object is used as there's only
one exported file.

Based on the history of the "exports" field that can be seen in [1], we
need to keep "main" around for at least as long as 12.x is specified as
a supported Node.js version (and to be on the safe side, 14.x as well).

--
1. https://nodejs.org/api/packages.html#exports
2. https://nodejs.org/api/packages.html#exports-sugar