Remove SAST from nightly nightly CI workflow
Don't run the reusable SAST workflow in the nightly workflow. The
motivation being that with pinned Action versions and a SAST scan on all
pushes to the `main` branch, analyzing the `main` branch nightly does
not add any value.

Additionally, the nightly run can cause interference with CodeQL because
it expects an analysis coming from the nightly workflow on the Pull
Request branch of the base happened to be the subject of a nightly scan.