Avoid running CodeQL in nightly continuous integration workflow
Adjust the CI setup to avoid running CodeQL in the nightly workflow to:

1. Avoid wasting CI time on rescanning commits that have already been
   scanned. Given that CodeQL is pinned, there's no reason to expect a
   different result when rescanning.
2. Avoid confusion on CodeQL parts due to no CodeQL scan from the
   nightly being found for non-`main` branches.

To avoid confusion with the naming of the reusable-sast.yml workflow
(because CodeQL is also SAST), the Gitleaks job has been merged into the
reusable-audit.yml workflow (I believe that fits in nicely).

To avoid running Gitleaks on the entire commit history twice tho (due to
the use of a matrix in nightly.yml), the reusable-audit.yml workflow has
been adjusted to accept a list of refs (as JSON) and create a matrix
internally instead. Note that this also sneakily introduces the
`fail-fast` option on the matrix for auditing with npm. This is to
ensure all audits are completed, given that they're mutually independent
of each other.

Lastly, to be able to test these changes, all references to the
reusable-audit.yml workflow have been updated to use the current branch.
This MUST be reverted prior to these changes landing in `main`.