Add markdownlint for linting MarkDown
Add the dependency markdownlint-cli to be able to conveniently run
markdownlint to lint MarkDown files. This is different from Prettier
because markdownlint aims to lint for mistakes and inconsistencies
rather than just formatting.

Running markdownlint is also included in the CI's lint check to ensure
continuous quality of MarkDown files.

A configuration file is included for completeness and convenience when
markdownlint starts to need to be configured.