Add setup for analyzing licenses with Fossa
Add a Fossa configuration file and an npm script to check the licenses
of the dependencies of the project using the Fossa CLI [1] (doing this
locally still requires a Fossa account, including an API key).

The project currently uses the Fossa GitHub integration, this setup
allows for more easily reproducing the license check locally. Also,
using the CLI allows for using a configuration file, being added here as
well.

--
1. https://github.com/fossas/fossa-cli