Configure incremental mode for mutation testing
Update the Stryker configuration file to include configuration for
incremental mode. Incremental mode is always enabled as the Mocha Runner
provides good support for rerunning the appropriate mutants when source
code or tests change.

Refactor the configuration slightly to re-use a constant for the path to
the mutation reports directory.