Add `audit:runtime` npm script
Add a new npm script called "audit:runtime" that is intended to be used
to audit runtime, or production, dependencies only. This is particularly
useful for end-users or when auditing on a version git ref rather than
development git ref (to see if a particular release had dependencies
with known vulnerabilities).

Also update the Contributing Guidelines in accordance with this
addition, and add a TODO to the reusable-audit.yml workflow to start
using this in the future (can't use it now because the `v0` ref, used in
nightly.yml, does not yet know about this script).