Drop "audit:runtime" script
Auditing runtime dependencies only is unnecessary as this is a library,
not only because there's currently no runtime dependencies but also
because it is generally not problematic for runtime dependencies being
depended upon to have known vulnerabilities as code is never run in an
interesting setting and users of the library can always upgrade to a
safe version because we would be using version ranges.

Correspondingly update the audit workflow in the CI and Contributing
Guidelines.

Note that this assessment excludes vulnerabilities in the malware
category.