Update eslint peerDependency range to be more exact
Like the `"engines"."node"` version range, only promise support for the
versions of ESLint already release. When ESLint v9 (or higher) is
released, this can be tested within the package and (if all is well) we
can update the version ranges to include `|| 9.x` (or higher)
followed by a new release.