Add licensee
Add licensee [1] as a dependency to be used as a license validation
tool. An initial configuration is included that defines an allowlist of
licenses based on the (permissive) licenses used by dependencies at this
point in the commit history. Note that 2 dependencies are not included
because they don't specify their license in a standard manner. As a
result, running the license checker (same command as with Fossa) will
fail (i.e. exit with a non-zero exit code).