Allow configuring allowed assignment for const declarations
Add a new option, `constAllowed`, that can be used to configure what
types of assignments are allowed for `const` declarations. The option is
also added to the documentation for this rule.

This is limited to `const` because assignments to `let` or `var`
declarations are not allowed at all (because they're mutable).

It is currently limited to allowing or not allowing only "Literals"
(e.g. `"foobar"` or `3.14`) as that's the only type of assignment
supported by this plugin at the moment.