Fix build issues on older Node.js version
The build script is only intended to work on the Node.js version
specified in `.nvmrc` or the Contributing Guidelines. Prior to upgrading
to rollup v3, it was also possible to build on all Node.js version back
to v12. However, rollup v3 is no longer backwards compatible with some
of the Node.js versions supported, and thus tested for, by this project.

First, update the run-compat-tests.js script such that it won't try to
build the JavaScript file in the container as the build might not work
in the given Node.js version. Instead, it will run the build command
before starting tests.

In particular, prior to this commit the script would report an error for
Node.js v12:

    Error: Cannot find module 'node:process'

which rollup uses internally (as of v3) but isn't supported by Node.js
v12.

Secondly, the compatibility tests in the CI suffer from a similar
problem. Here, this is solved by uploading the index.js file as an
artifact in the `build` job and downloading it in the `test-compat` job.