Add and configure eslint-plugin-markdown
Add eslint-plugin-markdown [1] as a devDependency and use it to lint
snippets in MarkDown files. For now, only JavaScript snippets are linted
as that is the only file type currently linted by ESLint.

In particular, the 1st ESLint override configures eslint-plugin-markdown
for all MarkDown files. The 2nd override configures ESLint for "files in
MarkDown files with the `.javascript` extension" (which is snippets with
the javascript annotation). And the `lind:md` npm script is updated to
run both markdownlint [2] and ESLint on MarkDown files only.

--
1. https://www.npmjs.com/package/eslint-plugin-markdown
2. https://github.com/DavidAnson/markdownlint