Refactor no-top-level-variables logic to reduce evaluated conditionals
Update the implementation of the no-top-level-variables rule to reduce
the number of conditional evaluated per declaration by extracting
conditionals that don't depend on the specifics of a given declaration
out of the declarations loop, as well as only checking the declaration's
`init.type` value as few times as absolutely necessary.

Note that the `isRequire` variable has been extracted separately for
`const` and non-`const` nodes. This was done because for `const` nodes
it is not necessary optionally chain `declaration.init?.type` as a
`const` must always be initialized. As a result, one more conditional is
avoided for `const` nodes.

Also, add `(at)types/estree` as an explicit devDependency now that the
source code imports type definitions from it.