Lint JSON files continuously
Update the "Check / Lint" CI job to run the `npm run lint:json` command
in order to lint JSON files continuously.

Also, update the job to always run all linters. This is to ensure all
lint violations are flagged on a single run, even if an early linter
detects a violation and exits with an error code.

Additionally, sneak in formatting fixes for `.eslintrc.js`.