Test again variables & side effects in top-level BlockStatements
Update the tests for both the "no-top-level-variables" and "no-top-level
-side-effect" rule to include a case where the variable or side-effect
(resp.) appears in a top-level BlockStatement. This should produce a
violation but was previously untested.

This untested case was found by mutation testing. This revealed that
line 5 of `lib/helpers.ts` could by changed as:

    -  while (scope.type === 'BlockStatement') {
    +  while (false) {

without it being flagged by a failing test.