Add and configure type-coverage
Add type-coverage [1] as a devDependency for vetting purposes and
configure it through CLI arguments (the altnerative is to do it in
`package.json`, which isn't desired). type-coverage is a vetting tool
"to check type coverage for typescript code", in particular reporting
on implicit `any`s and unchecked type casts.

Additionally, a new CI job is included to run vetting tools (just
type-coverage for now).

--
1. https://www.npmjs.com/package/type-coverage