Improve test suite for no-top-level-variables rule
Update the no-top-level-variables rule's test suite with additional test
cases based on the mutation report for the rules source code.

The two "valid" test cases added as well as the first two "invalid" test
cases added test partial configuration of the `constAllowed` option. The
fact that such tests were missing was uncovered as a result of
fallthrough mutation in the `allowedDeclaration` definition starting on
line 61.

The last "invalid" test case covers the scenario where there's a
top-level non-`require` function call assignment to a non-`const`
variable. This was detected as a mutation on line 84, which was only
tested for the positive case.