Improve test suite for no-top-level-side-effects rule
Update the no-top-level-side-effects rule's test suite with additional
test cases based on the mutation report for the rules source code.

The first "valid" test case covers IIFEs but with an arrow function
instead of a `function` function. The second "valid" test case covers
side-effects such as if-statements, for-statements, while-statements
and switch-statements but in a non-top-level context.

The "invalid" test cases are all about top-level assignments to
variables that are not either `module` or `exports`.