Avoid repeated test logic for compatibility tests
Convert the compatibility test snapshot data into a list of snapshots
and iterate over this list in the test suite.

Note, the test for the exit code was removed intentionally. It was
reconsidered because it's not the same for all snapshots (so it could've
been included in the snapshot data). However, I realized that the exit
code is actually a feature of ESLint and not this plugin, so there's no
point in asserting anything about it here.