Add and configure unimported
Add unimported [1] as a devDependency for vetting purposes and configure
it through a configuration file. unimported is a vetting tool to "Find
unused source files in javascript / typescript projects.",  It is
configured to look only at TypeScript code, not scripts or configuration
files, as that's the most important aspect.

The CI job for vetting automatically runs this new tool because it
simply invokes the "vet" command.

--
1. https://www.npmjs.com/package/unimported