Add and configure `depcheck`
Add `depcheck` [1] as a devDependency, run it as a vetting tool with the
`vet:deps` command (also ran by the `vet` command). Configure it with
the (intentionally explicitly referenced) `.depcheckrc.yml` file. This
configuration ensures depcheck correctly handles the TypeScript and
JavaScript code in this project, as well as ignoring some known false
positives.

--
1. https://www.npmjs.com/package/depcheck