Create asdf `.tool-versions` file
Create a `.tool-versions` file with currently untracked dependencies
(that we want to track). This file is first a configuration format for
asdf [1], but is also human-readable and thus can be used to know and
track what versions of dependencies _should_ be used.

Using the file with asdf currently requires the plugins:
- https://github.com/crazy-matt/asdf-actionlint
- https://github.com/luizm/asdf-shellcheck

All version are set to the latest available version. No violations are
found by these versions, so it was considered a sane version to choose.
The fact that versions are specified in .tool-versions is also now
referenced in the Contributing Guidelines.

--
1. https://asdf-vm.com/