Update "Check / Lint" and "Check / Vet"
Use `failure() || success()` for linting and vetting in CI instead of a
custom flag. With the custom flag steps following a failing step aren't
executed reliably. The new approach is preferred over the old approach,
with `always()`, as the old approach results in steps being executed
even if the job is cancelled or otherwise stopped.

Follows up on:
- 15087b860712e8ba2bb1117cbf54afc8d9b5ad66
- 24a8560af9ce9f0dab6ccfa0852c22d5b4a09e2e