Remove unnecessary thruthy check for IIFE evaluation
Update the no-top-level-side-effect rule implementation to remove an
unnecessary thruthiness check for IIFE expressions. This provides a
slight performance improvement as it's one less expression to evaluate
for top-level expressions (if this branch is reached).

The type system is able to infer the type of `node.expression.callee`
after the `node.expression.type === 'CallExpression'` evaluation. Hence,
checking if it is thruthy isn't necessary.