Configure audit exceptions using GitHub Security Advisory IDs
Update the `.nsprc` configuration file (for better-npm-audit [1]) to use
GitHub Security Advisory IDs to identify vulnerabilities to ignore. This
is instead of the previously used Node Security IDs which appear to be
not-so static based on experiences in different projects: the Node
Security ID for a given CVE/GHSA changed overnight.

--
1. https://www.npmjs.com/package/better-npm-audit