Enable verbose logging when linting in CI
To get some insight into what is being linted so this can be checked
without having to run things locally.

The "Lint MarkDown" step has been rewritten to allow for passing
additional argument to subcommands. In particular this is necessary to
enable debug logging for `npm run lint:md:code`.

As far as I was able to tell markdownlint-cli does not offer a verbose/
debug mode.