Apply SPDX identifiers
Add SPDX identifiers on the file level for source code in accordance
with <https://spdx.dev/ids/>

The `scripts/` are intentionally included here as they are meant to be
licensed using the same license as the source code.

Documentation is not included because I'm unsure how the existing license
applies to docs, and by extension I don't know what the original
author's (i.e. the author of, e.g., the commit de3072d) claims in this
regard could be.

Configuration files (CI or otherwise) are not included
because I'm unclear what it means to license a configuration file.