Replace Docker-based custom `:compat-all` script by nve
Add nve [1,2] as a devDependency to support running compatibility tests
on all applicable Node.js versions. This removes the (albeit optional)
dependency on Docker [3] from the project entirely. The CI is unaffected
as it just runs the compatibility suite for each applicable Node.js
version in a separate job (`check.yml:L238-L244`).

The new `test:compat-all` script definition duplicates the `test:compat`
script definition since using `npm run ...` is failing for Node.js v12 -
this is a minor drawback, but should be temporary.

The dependency on nve brings in the transitive dependency pako [4] that
is dual licensed under MIT and the Zlib license. Per the BlueOak council
[5] the Zlib license is a (not great) permissive license, hence it's
added here as an allowed dependency license.

--
1. https://www.npmjs.com/package/nve
2. https://github.com/ehmicky/nve
3. https://www.docker.com/
4. https://www.npmjs.com/package/pako
5. https://blueoakcouncil.org/list