Deduplicate the compatibility testing command
Avoid the repetation of the compatibility testing command between
`test:compat` and `test:compat-all` by simply invoking `test:compat`
(with the npm option `--ignore-scripts` to avoid unnecessarily
rebuilding). This is similar to the setup prior to [1], where the
duplication was introduced because invoking npm didn't work on older
(no-longer-supported) Node.js versions.

--
1. 5e43ea163061b5f5a1ea2b26c2f5a71247cc6d11