Bump transitive dependency optionator & ignore word-wrap vulnerability (#497)
Update the version of optionator used transitively through eslint-v7 in
order to resolve the transitive dependency on the vulnerable and
unmaintained dependency word-wrap. This was achieved by running `npm
audit fix`, thus includes additional transtivie dependency bumps as
well.

Also ignore the same vulnerability in word-wrap because it's still
present in the dependency tree of this project through eslint-v6. It's
ignored with the same reason as the vulnerability in glob-parent.