Update dependabot.yml
Configure two `groups` per [1] to streamline dependency updates by
Dependabot. The first group is aimed at the Stryker dependencies, they
are always released together and in the past this project has
experienced a lack of updates in this area (seemingly because of inter-
connected dependency requirements). The seconds is aimed at the
TypeScript ESLint dependencies, while there haven't been any problems in
the past, this will reduce the number of updates received as these
updates are also always released together.

--
1. https://github.blog/changelog/2023-06-30-grouped-version-updates-for-dependabot-public-beta/