Add basic reproducible build job (#668)
Create a new CI job that verifies builds are reproducible for this
project. This job depends on the build job as the build shouldn't
be expected to be reproducible if it doesn't work to begin with.

The StepSecurity harden-runner policy is based on the web interface
recommendations as well as the pre-existing policy for the related
`build` job.