Disable lifecycle scripts by default
Update the repository configuration to disable lifecycle scripts (such
as installation scripts) by default. None of this project's current
dependencies require such scripts, so from that perspective it's fine.
Perhaps the most challenging thing is that the `prepublishOnly` script
no longer runs when publishing. To mitigate this all invocations of `npm
publish` in the project explicitly override `--ignore-scripts` to enable
lifecycle hooks, and the release guidelines also recommend explicitly
running it to reduce the risk of user error.