Test assignment of IIFE and `new`, report on the latter
Update the `no-top-level-side-effects` rule to report on the presence of
side effects in variable declarations in the form of `new` statements.

Also test that the `no-top-level-side-effects` rule reports on the
presence of side effects in variable declarations in the form IIFEs.
These aren't supposed to be allowed through the `allowIIFE` option
because the use case for that is strictly in top-level expression
statements.