Optionally allow creating `Symbol` as top-level side effect
Adjust the reporting for `no-top-level-side-effects` to optionally allow
calling `Symbol()` at the top level. Initializing symbols at the top
level is the only way to create symbols at the top-level, and can be
viewed as similar to initializing to a literal. This is allowed by
default, but can be disabled.