Find alternative for lifecycle hooks on own scripts
Update the `vet:package.json` script to directly run the old lifecycle
hook.

Update the `test:compat` and `test:compat-all` scripts to omit the
lifecycle hook altogether and require manually calling `npm run build`.
This is necessary because inlining the old lifecycle hook would not work
when trying to run the compatibility using older Node.js version because
we don't want to put the requirement on the build that it works on other
Node.js versions.