Optionally allow top-level `const` template literals
Adjust the reporting for `no-top-level-variables` to optionally allow
assigning template literals to top level `const`s. This is allowed by
default because strings (one possibility of literals) are allowed by
default as well. The fact that interpolation can be seen as a side
effect should be handled by the `no-top-level-side-effects` rule instead