Rework the `no-top-level-variables` rule
Re-implement the `no-top-level-variables` rule to focus it on variables.
In particular, this aims to remove checking for side-effects (e.g.
function calls) from the rule and instead focus on the presence of
mutable top-level variables.

The default configuration is geared towards minimizing mutable top-level
variables while being sensible. For example, an ArrowFunctionExpression
is theoretically mutable but is not typically used in that way (also,
a FunctionStatement is theoretically mutable as well).

Additionally, the configuration allows for a more lenient setup, e.g. by
allowing top-level variables to be objects or arrays, which are mutable
(and typically used so).

Lastly, the configuration also allows for using `var` and `let` if that
is necessary for historical or compatibility reasons. But they're
completely disallowed by default because they're always mutable.