Allow user to disallow use of `require`
Update the `no-top-level-side-effects` rule with an option to disallow
the use of `require`. This can be useful in non-CJS settings where for
some reason `require` is an actual function. In this case you don't
actually want to allow use of `require` at the top level.

To support this change, some refactoring has been carried out as well in
order to reduce duplication.