Disable lifecycle scripts by default (#735)
Update the repository configuration to disable lifecycle scripts (such
as installation scripts) by default. None of this project's current
dependencies require such scripts, so from that perspective it's fine.
Perhaps the most challenging thing is that the `prepublishOnly` script
no longer runs when publishing. To mitigate this all invocations of `npm
publish` in the project explicitly override `--ignore-scripts` to enable
lifecycle hooks, and the release guidelines also recommend explicitly
running it to reduce the risk of user error.

As a result, `pre*` lifecycle hooks defined in this project's manifest
also don't work anymore. As such, update the `vet:package.json` script
to directly run the old lifecycle hook. Update the `test:compat` and
`test:compat-all` scripts to omit the lifecycle hook altogether and
require manually calling `npm run build`. This is necessary because
inlining the old lifecycle hook would not work when trying to run the
compatibility using older Node.js version because we don't want to put
the requirement on the build that it works on other Node.js versions.