Replace `allowExports` option by `commonjs`
Update the `no-top-level-side-effects` rule to omit the `allowExports`
option and instead adopt the `commonjs` option. This new option allows
managing whether all of `require`, `module.exports`, and `exports` are
allowed. When set to true, they're all allowed (`module.exports` and
`exports` are handled as before, `require` is injected into the list of
allowed functions to call).

This option is off by default to encourage the use of ESM and to avoid
false negatives in ESM code.