Add option to allow derivations as `top-level-side-effects`
Update the `top-level-side-effects` rule to include an option that
enables users to allow top-level declarations to perform side-effect
free operations on values and variables. This is disabled by default
because the act of performing the derivation is a side effect. However,
because there's no side effects to the derivations and because it can
be useful (e.g. `const timeout = 24 * 60 * 60` might be easier to read),
users are given the ability to allow these in their code base if they so
choose.