Add option to allow derivations as `top-level-side-effects` (#785)
Update the `top-level-side-effects` rule to include an option that
enables users to allow top-level declarations to perform side-effect
free operations on values and variables. This is disabled by default
because the act of performing the derivation is a side effect. However,
because there's no side effects to the derivations and because it can
be useful (e.g. `const timeout = 24 * 60 * 60` might be easier to read),
users are given the ability to allow these in their code base if they so
choose.

Also refactor the implementation so that it can recursively check
operands of binary, logical, and unary expression (if derivations are
allowed). This will catch other side effects within the subexpression
tree (e.g. `f()` in `x + f()`).