Audit only when necessary (#857)
* Audit only when necessary

Update the `audit.yml` workflow to only trigger for Pull Requests when
there is a chance for a new finding as a result of the change. It is
only possible to introduce a new vulnerability when the dependencies
change (covered by `package-lock.json`) or when the auditing config has
changed (covered by `.nsprc`). The motivation for this is to reduce
unexpected CI failures.

We also always run this workflow when it is changed as a sanity check.

* Extract secret job from `reusable-audit.yml`

Because we want to trigger the secret job for most triggers but audit
npm dependencies for fewer triggers.