One case per test for `no-top-level-variables`
Update the test suite for `no-top-level-variables` such that there's
only one case per test. To keep some sort of grouping that was there
with the previous approach, related tests are grouped together in a sub-
array.

Besides that changes this also abstracts the options away for improved
readability within tests and also fewer changes if the options need to
change.