Add a preset strict configuration (for ESLint v9 configurations)
Create and export a preset strict configuration object to make adopting
this plugin to its full effect easier. The goal of this prest config is
to disallow as much as possible. Like the recommended config, it does
set `commonjs: true` for the `no-top-level-side-effects` rule to avoid
having to mess with the configuration for CommonJS-based projects.