Automatically determine `commonjs` option value if unspecified
Update the implementation of the `no-top-level-side-effects` rule to
determine if the given node is from a CJS or ESM piece of code by
looking at the `sourceType` of the root `Program` node [1].

This functionality is only enabled when the `commonjs` option is not
specified. If it is specified it always takes precedence.

--
1. https://github.com/estree/estree/blob/3851d4a6eae5e5473371893959b88b62007469e8/es2015.md#programs