Rewrite how we determine if a node is at the top level
Update the `isTopLevel` to consider more statements and expression to be
part of the top level, similar to `BlockStatement`, so that expression
are "automatically" considered top-level if they appear in said
expressions or statements.

This change was suggested by a contributor who deservers full credit for
this idea - see the commit's co-author - I merely did the engineering
work of refactoring to code.

Co-authored-by: Yurui Zhang <yuruiology@gmail.com>