Explicitly test extra allowed option values for `no-top-level-variables`
Extend the test suite for the `no-top-level-variables` rule with tests
for each "additional expression type" that can be configured. This is
meant as capturing what is currently allowed (now that we have test
cases for them I feel like reconsidering if they shouldn't always be
allowed).

Notably, the `YieldExpression` is now removed from the documentation.
This is because it is not actually possible to put a `yield` at the top
level since it's a reserved keyword that's only allowed in generators.
Importantly, the option is still allowed in the rule configuration as
removing it would be a breaking change - I captured this with a test
case that configures this option just to ensure it's not rejected.