Allow three more expression in `no-top-level-variables`
Update the `no-top-level-variables` rule to always allow the expressions
`ImportExpression`, `SequenceExpression`, and `ThisExpression` in top
level variables. This has been reflected in the implementation, tests,
and documentation.

These expression types are still included in the enum for optionally
allowed expressions so as to not break existing configurations that
allow them (otherwise ESLint would complain about unknown values). The
tests have also been updated to be explicit about this.