Switch `eslint-plugin-markdown` to `(at)eslint/markdown`
Following [1] this replaces `eslint-plugin-markdown` by
`(at)eslint/markdown` for linting code snippets in MarkDown for this
project. The motivation for this change is primarily to switch to a
modern and "first-party" plugin for this purpose.

Notably, this commit also rewrites the ESLint configuration from CJS to
ESM, this is because `(at)eslint/markdown` is only available as ESM and
so cannot be imported from CJS.

--
1. https://eslint.org/blog/2024/10/eslint-json-markdown-support/