Use `depreman` to manage deprecations (#1234)
* Use depreman to manage deprecations

It's an early prototype by me to attempt to manage deprecations. Using
it here to dogfeed it on my own projects and understand what needs to
change or be supported as well as just to start managing the deprecated
dependencies of this project.

First thing I noticed: while `"transitive": true` is desired for both
eslint rules, it actually isn't for mocha and publint where I'd rather
ignore a specific transitive dependency only, because I don't know
intent to ignore other transitive deprecated packages that may come up
in the future for these.

* Update licensee configuration

Explicitly allow depreman which is licensed under the `AGPL-3.0-only`,
which is fine for a devDependency but would not desirable for
"production" dependencies (given this project is permissively licensed).