Upgrade depreman (#1236)
Update depreman to v0.2.0 with improved configuration that fits the
needs of this project.

This also changes how depreman is used, instead of vetting it is now
considered as an audit process. This aligns it with expectations: vet
command should always pass and audit commands may fail (because both
vulnerabilities and deprecations may arise without changes to the code
base).