Re-organize development commands
Attempt to avoid the arbitrary divide between vetting, linting, and
other checks by just naming them all "checks" (and referring to it as
analysis in the Contributing Guidelines). This is inspired by the CI
workflow called "Check" too.

The only potentially seemingly related categorization left now is checks
and audits. The division here is that "checks" should always pass (are
invariant w.r.t. the commit history) while audits don't (because
packages may be deprecated or report vulnerability at any time).

On top of that, this includes some minor reformatting of the
Contributing Guidelines, and also introduces `npm-run-all2` as a
convenience to run all checks (and audits).