Harden GitHub Actions workflows (#1275)
Update all GitHub Actions workflows following an analysis by zizmor
(https://github.com/woodruffw/zizmor). In particular, this avoids
persisting git credentials when the job does not need it.

Zizmor has two more concerns. First, overly permissive `permissions`,
which not addressed because I think this is okay, the project is
entirely open so I don't see a risk of an attacker reading anything.
Second, the use of `pull_request_target` in the labeler workflow, which
was not addressed because this job does not run any untrusted code and
it's the only way to get the labeler action to work properly.