Try out `eslint-plugin-depend`
Add `eslint-plugin-depend` as a dependency and use it with ESLint to
suggest alternatives to various dependencies with the goal of preventing
dependency tree bloat.

As of this commit the plugin does not detect any problems, but the idea
is that it would prevent problems from being introduced later OR being
detected as the plugin is updated.

As a testing example, adding an extra imports as

    -
    + import fs from "fs-extra";

will cause the plugin to produce a lint error.