Optionally allow adding properties to function declarations
Add an option to the `no-top-level-side-effects` rule that allows for
adding a property to a function declaration. This supports the pattern
where a function declaration doubles as an object, as seen in e.g.,
functional React components.

This is necessary because there is no other way for functions to be
assigned (static) properties. Hence, this is the only way to support
this pattern. To limit misuse of this option, the implementation checks
that the assignment occurs on a locally defined top-level function
declaration (rejecting e.g. imported functions or locally declared
objects).