Support allowing calls to member expressions
Add support to the no-top-level-side-effects rule's `allowedCalls`
option to allow calls to member expressions such as `console.log` or
`Symbol.for`.

The `expr.property` is cast to an `Identifier` in the `isCallTo`
function because as a result of the `MemberExpression` being not
computed it must be that the property is an identifier.