Replace `ts-node` by `tsx`
This replaces `ts-node` by `tsx` for "running" TypeScript code directly.
Unlike `ts-node`, `tsx` (just) works with Node.js v24. On the downside,
`tsx` struggles a bit with Node.js v18 (which `ts-node` did not). But,
this can be fixed by just hotpatching the Mocha configuration. It's not
pretty but it gets the job done, and it's nicer to have an ugly solution
for an old Node.js version than for a new Node.js version.