Support allowing top-level calls of membership expression (#1475)
Add support to the no-top-level-side-effects rule's `allowedCalls`
option to allow calls to member expressions such as `console.log` or
`Symbol.for`.


Implementation note:
The `expr.property` in the `isCallTo` function is cast to an
`Identifier` because as a result of the `MemberExpression` being
not computed it must be that the property is an identifier.