Add support for Node.js v24 (#1474)
This updates the package to support using the plugin on Node.js v24 officially.
The tests are now also run against that version in CI. To support this, the
compatibility test suite is rewritten to vanilla JavaScript.

This rewriting was done because it was the simplest way I could find to run
these tests on all supported Node.js versions. It appears `ts-node` does not
work on Node.js v24 (and I do not expect that to change given it's lack of
updates in a long time). I tried `tsx` as well, which works on Node.js v20 to
v24 and, given an alternative Mocha config, can also work on Node.js v18. This
would be fine except that it gives an incorrect test coverage result on Node.js
v22 (the current version used for development). With `tsx` the test coverage
report is correct on Node.js v24, however we can't upgrade to that version of
Node.js yet because the mutation testing framework does not yet work fully on
Node.js v24.

Hence, once the development version of Node.js is bumped to v24, the compat test
suite may be reverted back to TypeScript. However, it does not make sense to
block the support of Node.js v24 by the plugin until the development version of
the project can be upgraded too. Therefor, these changes were accepted as is.