Convert compatibility tests back to TypeScript
Partially revert b2ffec5686b83fd64a7148eb069a399d6a76d316, changing the
compatibility test suite back to TypeScript. This is enabled by
switching from `ts-node` to `tsx`, which works on Node.js v24. However,
the Mocha configuration requires a slight change on Node.js v18, while
not ideal, I consider this an OK trade-off since this should be uncommon
(locally) in practice, and is automated in CI. Still, this has been
added as a note to the Contributing Guidelines for completeness.